refactor: chronometer component into its own file for better modularity and maintainability
